% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pop_evolution.R
\name{get_pop_evolution}
\alias{get_pop_evolution}
\title{Calculate and plot population evolution}
\usage{
get_pop_evolution(
  df,
  species,
  locality,
  n = 100,
  years = 10,
  colours = "#000000"
)
}
\arguments{
\item{df}{A dataframe with population dynamic parameters. It must contain (at least) the following columns:
\enumerate{
\item \code{locality}
\item \code{species}: vernacular name.
\item \code{lifestage}
\item \code{reproduction}: probability to reproduce within a year. Real number
equal or above 0.
\item \code{survival}: probability to survive within a year. Real number
between 0 (the animal dies for sure) and 1 (the animal survives for sure).
}}

\item{species}{Character with the species name. One of the species in column
\code{species} of \code{df}.}

\item{locality}{Character with the locality name. One of the localities in
column \code{locality} of \code{df}.}

\item{n}{Number of individuals per  life stage class. Default: 100
individuals for each class.}

\item{years}{Number of years for the prediction. Default: 10 years.}

\item{colours:}{vector with hex colour codes, one per life stage. If one
value is passed, all life stage categories have same color in the returned
line plot. Default: black for all life stages.}
}
\value{
A ggplot2 line plot with years on the x-axis, number of animals on
the y-axis and the lifestage as colour and label. The lambda value is shown
in the subtitle.
}
\description{
Calculate and plot population evolution
}
\examples{
library(fisrutils)
get_pop_evolution(pop_dyn, species = "wild boar", locality = "Flanders")
}
