% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pop_dyn}
\alias{pop_dyn}
\title{Sample of Population Dynamic Parameters}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 10 rows and 5 columns.
}
\source{
\url{https://github.com/tdwg/camtrap-dp/tree/ad0278ef86ef518dacfb306c598dce97667cfb81/example}
}
\usage{
pop_dyn
}
\description{
A tibble (dataframe) containing a sample of Population Dynamic Parameters.
}
\details{
Columns:
\enumerate{
\item \code{locality}
\item \code{species}: vernacular name.
\item \code{lifestage}
\item \code{reproduction}: probability to reproduce within a year. Real number
equal or above 0.
\item \code{survival}: probability to survive within a year. Real number
between 0 (the animal dies for sure) and 1 (the animal survives for sure).
}
}
\examples{
\dontrun{
# pop_dyn.rda was created with the code below.
pop_dyn <- read.csv2(
  system.file(
    "tests/testthat/data/",
    "pop_dyn.csv",
    package = "fisrutils"
  ),
  sep = ";"
)
save(pop_dyn, file = "data/pop_dyn.rda")
}
}
\concept{sample data}
\keyword{datasets}
